support direct connections to acraserver from clients with static client id
